jestが失敗するようになっていたので修正、backendのテストをtestフォルダのみで実行するように修正
