+++ packages/frontend/src/app/(authenticated)/context/NoteProvider.tsx
@@ -23,13 +23,17 @@ export const NoteProvider: React.FC<{ children: React.ReactNode }> = ({ children
 
     // Noteを取得
     const fetchNotes = async () => {
-        const response = await fetch("/api/notes", {
-            method: "GET",
-            credentials: "include"
-        });
-        if (!response.ok) throw new Error("Failed to fetch notes");
-        const data = await response.json();
-        setNotes(data);
+        try {
+            const response = await fetch("/api/notes", {
+                method: "GET",
+                credentials: "include"
+            });
+            if (!response.ok) throw new Error("Failed to fetch notes");
+            const data = await response.json();
+            setNotes(data);
+        } catch (error) {
+            console.error("Error fetching notes:", error);
+        }
     };
 
     useEffect(() => { fetchNotes(); }, []);
