+++ packages/frontend/test/app/(authenticated)/components/Note.test.tsx
@@ -1,32 +1,16 @@
 import React, { act } from "react";
 import { render, screen, fireEvent, within, waitFor } from "@testing-library/react";
 import "@testing-library/jest-dom";
-import Note from "../../src/app/(authenticated)/components/Note";
+import Note from "@/app/(authenticated)/components/Note";
 // ラベルコンテキストのモック
 const mockLabels = [
     { id: "label1", labelname: "仕事" },
     { id: "label2", labelname: "プライベート" },
 ];
 
-type Column = {
-    id: number;
-    name: string;
-    order?: number;
-}
-
-type RowCell = {
-    id: number;
-    rowIndex: number;
-    value: string;
-    columnId?: number;
-}
-
-const mockColumns: Column[] = [{ id: 1, name: "テストカラム1", order: 1 }];
-const mockRowCells: RowCell[][] = [[{ id: 1, rowIndex: 0, value: "テストセル１", columnId: 1 }], [{ id: 1, rowIndex: 1, value: "テストセル２", columnId: 1 }]];
-
-jest.mock("@/context/LabelProvider", () => {
+jest.mock("@/app/(authenticated)/context/LabelProvider", () => {
     return {
-        ...jest.requireActual("@/context/LabelProvider"),
+        ...jest.requireActual("@/app/(authenticated)/context/LabelProvider"),
         useLabelContext: () => ({
             labels: mockLabels,
         }),
@@ -36,10 +20,9 @@ jest.mock("@/context/LabelProvider", () => {
 
 import { LabelProvider } from "@/app/(authenticated)/context/LabelProvider";
 import { NoteProvider } from "@/app/(authenticated)/context/NoteProvider";
-import TableNote from "@/app/(authenticated)/components/TableNote";
 
 
-describe("TableNote", () => {
+describe("Note", () => {
     const mockOnSave = jest.fn();
     const mockOnDelete = jest.fn();
 
@@ -51,16 +34,17 @@ describe("TableNote", () => {
         }
     });
 
-    it("openがfalseのとき、タイトルと作成日、更新日が表示される", () => {
+    it("openがfalseのとき、タイトルとcontent、作成日、更新日が表示される", () => {
         render(
             <NoteProvider>
                 <LabelProvider>
-                    <TableNote id={"testid111"} title={"テストノート"} label_id={""} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} columns={mockColumns} rowCells={mockRowCells} onSave={mockOnSave} onDelete={mockOnDelete} />
+                    <Note id={"testid111"} title={"テストノート"} content={"テストノートcontent"} label_id={""} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} onSave={mockOnSave} onDelete={mockOnDelete} />
                 </LabelProvider>
             </NoteProvider>
         );
 
         expect(screen.getByText("テストノート")).toBeVisible();
+        expect(screen.getByText("テストノートcontent")).toBeVisible();
         expect(screen.getByText("作成日: 2025/07/05")).toBeVisible();
         expect(screen.getByText("更新日: 2025/07/05")).toBeVisible();
     })
@@ -69,7 +53,7 @@ describe("TableNote", () => {
         render(
             <NoteProvider>
                 <LabelProvider>
-                    <TableNote id={"testid111"} title={"テストノート"} label_id={""} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} columns={mockColumns} rowCells={mockRowCells} onSave={mockOnSave} onDelete={mockOnDelete} />
+                    <Note id={"testid111"} title={"テストノート"} content={"テストノートcontent"} label_id={""} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} onSave={mockOnSave} onDelete={mockOnDelete} />
                 </LabelProvider>
             </NoteProvider>
         );
@@ -87,7 +71,7 @@ describe("TableNote", () => {
         render(
             <NoteProvider>
                 <LabelProvider>
-                    <TableNote id={"testid111"} title={"テストノートタイトル"} label_id={""} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} columns={mockColumns} rowCells={mockRowCells} onSave={mockOnSave} onDelete={mockOnDelete} />
+                    <Note id={"testid111"} title={"テストノートタイトル"} content={"テストノートcontent"} label_id={""} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} onSave={mockOnSave} onDelete={mockOnDelete} />
                 </LabelProvider>
             </NoteProvider>
         );
@@ -110,11 +94,11 @@ describe("TableNote", () => {
 
     });
 
-    it("編集モード時、columnを編集できる", async () => {
+    it("編集モード時、contentを編集できる", async () => {
         render(
             <NoteProvider>
                 <LabelProvider>
-                    <TableNote id={"testid111"} title={"テストノートタイトル"} label_id={""} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} columns={mockColumns} rowCells={mockRowCells} onSave={mockOnSave} onDelete={mockOnDelete} />
+                    <Note id={"testid111"} title={"テストノートタイトル"} content={"テストノートcontent"} label_id={""} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} onSave={mockOnSave} onDelete={mockOnDelete} />
                 </LabelProvider>
             </NoteProvider>
         );
@@ -127,67 +111,13 @@ describe("TableNote", () => {
             fireEvent.click(screen.getByText("編集"));
         })
 
-        const contentInput = screen.getByDisplayValue("テストカラム1") as HTMLInputElement;
+        const contentInput = screen.getByDisplayValue("テストノートcontent") as HTMLInputElement;
 
         await act(async () => {
-            fireEvent.change(contentInput, { target: { value: "新しいカラム" } });
+            fireEvent.change(contentInput, { target: { value: "新しいcontent" } });
         });
 
-        expect(contentInput.value).toBe("新しいカラム");
-
-    });
-
-    it("カラムの追加ができる", async () => {
-        render(
-            <NoteProvider>
-                <LabelProvider>
-                    <TableNote id={"testid111"} title={"テストノートタイトル"} label_id={""} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} columns={mockColumns} rowCells={mockRowCells} onSave={mockOnSave} onDelete={mockOnDelete} />
-                </LabelProvider>
-            </NoteProvider>
-        );
-
-        await act(async () => {
-            fireEvent.click(screen.getByText("テストノートタイトル"));
-        })
-
-        await act(async () => {
-            fireEvent.click(screen.getByText("編集"));
-        })
-        const addColumnButton = screen.getByTestId("addColumnIcon");
-
-        await act(async () => {
-            fireEvent.click(addColumnButton);
-        });
-
-        await waitFor(() => {
-            expect(screen.getByDisplayValue("カラム2")).toBeInTheDocument();
-        });
-    });
-
-    it("編集モード時、rowCellを編集できる", async () => {
-        render(
-            <NoteProvider>
-                <LabelProvider>
-                    <TableNote id={"testid111"} title={"テストノートタイトル"} label_id={""} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} columns={mockColumns} rowCells={mockRowCells} onSave={mockOnSave} onDelete={mockOnDelete} />
-                </LabelProvider>
-            </NoteProvider>
-        );
-
-        await act(async () => {
-            fireEvent.click(screen.getByText("テストノートタイトル"));
-        })
-
-        await act(async () => {
-            fireEvent.click(screen.getByText("編集"));
-        })
-
-        const contentInput = screen.getByDisplayValue("テストセル１") as HTMLInputElement;
-
-        await act(async () => {
-            fireEvent.change(contentInput, { target: { value: "新しいセル" } });
-        });
-
-        expect(contentInput.value).toBe("新しいセル");
+        expect(contentInput.value).toBe("新しいcontent");
 
     });
 
@@ -195,7 +125,7 @@ describe("TableNote", () => {
         render(
             <NoteProvider>
                 <LabelProvider>
-                    <TableNote id={"testid111"} title={"テストノートタイトル"} label_id={"label1"} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} columns={mockColumns} rowCells={mockRowCells} onSave={mockOnSave} onDelete={mockOnDelete} />
+                    <Note id={"testid111"} title={"テストノートタイトル"} content={"テストノートcontent"} label_id={"label1"} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} onSave={mockOnSave} onDelete={mockOnDelete} />
                 </LabelProvider>
             </NoteProvider>
         );
@@ -225,7 +155,7 @@ describe("TableNote", () => {
             <>
                 <NoteProvider>
                     <LabelProvider>
-                        <TableNote id={"testid111"} title={"テストノートタイトル"} label_id={"label1"} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} columns={mockColumns} rowCells={mockRowCells} onSave={mockOnSave} onDelete={mockOnDelete} />
+                        <Note id={"testid111"} title={"テストノートタイトル"} content={"テストノートcontent"} label_id={"label1"} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} onSave={mockOnSave} onDelete={mockOnDelete} />
                     </LabelProvider>
                 </NoteProvider>
                 <input data-testid="dummy-input" placeholder="ダミー入力" />
@@ -255,7 +185,7 @@ describe("TableNote", () => {
         render(
             <NoteProvider>
                 <LabelProvider>
-                    <TableNote id={"testid111"} title={"テストノートタイトル"} label_id={""} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} columns={mockColumns} rowCells={mockRowCells} onSave={mockOnSave} onDelete={mockOnDelete} />
+                    <Note id={"testid111"} title={"テストノートタイトル"} content={"テストノートcontent"} label_id={""} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={false} onSave={mockOnSave} onDelete={mockOnDelete} />
                 </LabelProvider>
             </NoteProvider>
         );
@@ -289,7 +219,7 @@ describe("TableNote", () => {
         render(
             <NoteProvider>
                 <LabelProvider>
-                    <TableNote id={"testid111"} title={"テストノートタイトル"} label_id={""} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={true} columns={mockColumns} rowCells={mockRowCells} onSave={mockOnSave} onDelete={mockOnDelete} />
+                    <Note id={"testid111"} title={"テストノートタイトル"} content={"テストノートcontent"} label_id={""} createdate="2025-07-05 05:33:05.864" updatedate="2025-07-05 05:33:05.864" is_locked={true} onSave={mockOnSave} onDelete={mockOnDelete} />
                 </LabelProvider>
             </NoteProvider>
         );
