+++ packages/frontend/test/app/(authenticated)/components/TrashNote.test.tsx
@@ -1,13 +1,13 @@
 import React, { act } from "react";
 import { render, screen, fireEvent } from "@testing-library/react";
 import '@testing-library/jest-dom';
-import TrashNote from "../../src/app/(authenticated)/components/TrashNote";
+import TrashNote from "@/app/(authenticated)/components/TrashNote";
 import { LabelProvider } from "@/app/(authenticated)/context/LabelProvider";
 import { NoteProvider } from "@/app/(authenticated)/context/NoteProvider";
 
-jest.mock("@/context/LabelProvider", () => {
+jest.mock("@/app/(authenticated)/context/LabelProvider", () => {
     return {
-        ...jest.requireActual("@/context/LabelProvider"),
+        ...jest.requireActual("@/app/(authenticated)/context/LabelProvider"),
         useLabelContext: () => ({
             labels: mockLabels,
         }),
