+++ packages/frontend/test/app/(authenticated)/layout.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen, fireEvent, waitFor } from "@testing-library/react";
-import RootLayout from "../../src/app/(authenticated)/layout";
+import AuthenticatedLayout from "@/app/(authenticated)/layout";
 import "@testing-library/jest-dom";
 
 // fetch モック
@@ -11,12 +11,20 @@ global.fetch = jest.fn(() =>
     })
 ) as jest.Mock;
 
+const mockedUseRouter = jest.fn();
+
+jest.mock("next/navigation", () => ({
+    useRouter: () => mockedUseRouter(),
+    usePathname: jest.fn().mockReturnValue("/"),
+    useServerInsertedHTML: jest.fn(),
+}));
+
 describe("RootLayout", () => {
     it("renders logo, nav, and children", async () => {
         render(
-            <RootLayout>
+            <AuthenticatedLayout>
                 <div>Child Content</div>
-            </RootLayout>
+            </AuthenticatedLayout>
         );
 
         // ロゴ
@@ -35,28 +43,31 @@ describe("RootLayout", () => {
 
     it("カラーテーマ切り替え", () => {
         render(
-            <RootLayout>
+            <AuthenticatedLayout>
                 <div>Toggle Test</div>
-            </RootLayout>
+            </AuthenticatedLayout>
         );
 
-        const toggleButton = screen.getByTestId("togglecolormode");
-        expect(toggleButton).toBeInTheDocument();
+        const toggleButtons = screen.getAllByTestId("togglecolormode");
+        expect(toggleButtons.length).toBeGreaterThan(0);
+
+        // 最初のトグルボタンを使う
+        const toggleButton = toggleButtons[0];
 
         // 初期状態はlight → iconは <LightModeOutlinedIcon />
-        expect(toggleButton.querySelector("svg")).toBeInTheDocument();
+        expect(toggleButton.querySelector("svg[data-testid='LightModeOutlinedIcon']")).toBeInTheDocument();
 
         fireEvent.click(toggleButton);
 
         // ダークモードのアイコンに変化（Brightness2OutlinedIcon）
-        expect(toggleButton.querySelector("svg")).toBeInTheDocument();
+        expect(toggleButton.querySelector("svg[data-testid='Brightness2OutlinedIcon']")).toBeInTheDocument();
     });
 
     it("ラベルアイテムがクリックされたとき、ダイアログが表示される", async () => {
         render(
-            <RootLayout>
+            <AuthenticatedLayout>
                 <div>Label Dialog Test</div>
-            </RootLayout>
+            </AuthenticatedLayout>
         );
 
         const labelButton = screen.getByText("ラベル");
